requirements.txt
asgiref==3.2.7
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.7
django-ckeditor==5.9.0
django-cors-headers==3.3.0
django-js-asset==1.2.2
django-rest-knox==4.1.0
django-storages==1.9.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.10
marshmallow==3.8.0
mysqlclient==1.4.6
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
